Move files and cleanup code.
